Initial commit with some infrastructure.
1) autoload/, ftplugin/, plugin/ , pygerrit/ directories.
- autoload/ declares functions that are namespaced and referenceable
from the plugin's various other places. The gv.vim is the file that
declares some of the utility functions and does the actual bindings to
the python code.
- ftplugin/python/ contains vim code that binds keys to functions that
call auto load functions that call the python code.
- plugin/ contains the code that is specific to the plugin environement.
Like disabling/enabling the plugin, chacking for what python you have.
All the Gnarly stuff.
- pygerrit/ contains all the python bindings that pull things from
gerrit and displays it onto buffers.