Making things work!
Active right now!. the GvShowStatus command.
You need an active SSH Tunnel on port 8081 to gerrit.
Also a config/config.json file with username and password fields
to access gerrit needs to be in the plugin root directory.

Next up: Displaying files specific to the changes. Along with
a quick fix window to navigate through comments.